% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{At_Age_TS_Variables}
\alias{At_Age_TS_Variables}
\title{At-Age Time Series Variables}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
At_Age_TS_Variables
}
\description{
At-Age Time Series Variables
}
\keyword{datasets}
