% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demo}
\alias{demo}
\title{Run an example MSE}
\usage{
demo()
}
\description{
Run an example MSE using three data-limited management procedures from
\code{DLMtool} and one stock assessment model from \code{SAMtool}
}
\examples{
\dontrun{
demo()
}
}
