% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_at_Age}
\alias{get_at_Age}
\alias{get_at_Age.Hist}
\alias{get_at_Age.list}
\alias{get_at_Age.MSE}
\alias{get_at_Age.multiHist}
\alias{get_at_Age.MMSE}
\title{Create a data.frame with at-age schedules by simulation and year}
\usage{
get_at_Age(x, model = "Model 1", ...)

\method{get_at_Age}{Hist}(x, model = "Model 1", ...)

\method{get_at_Age}{list}(x, model = "Model 1", ...)

\method{get_at_Age}{MSE}(x, model = "Model 1", ...)

\method{get_at_Age}{multiHist}(x, model = "Model 1", ...)

\method{get_at_Age}{MMSE}(x, model = "Model 1", ...)
}
\arguments{
\item{x}{An object of class \code{Hist}, \code{MSE}, or a list of \code{Hist} or \code{MSE} objects}

\item{model}{An optional name for the model.
If \code{x} is a list of objects, \code{model} will be taken from \code{names(x)}.
If \code{names(x)} is \code{NULL}, \code{model} will be given sequential numerical values (e.g., \verb{Model 1}, \verb{Model 2}, ...)}

\item{...}{additional arguments}
}
\value{
A data.frame
}
\description{
Note that the Selectivity and Retention curves in these plots are from the operating model. If an MP changes the
selectivity/retention, this is not shown in these plots.
}
