% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_LifeHistory}
\alias{get_LifeHistory}
\alias{get_LifeHistory.Hist}
\alias{get_LifeHistory.list}
\alias{get_LifeHistory.MSE}
\title{Get Life History Parameters}
\usage{
get_LifeHistory(x, model = "Model 1")

\method{get_LifeHistory}{Hist}(x, model = "Model 1")

\method{get_LifeHistory}{list}(x, model = "Model 1")

\method{get_LifeHistory}{MSE}(x, model = "Model 1")
}
\arguments{
\item{x}{An object of class \code{Hist}, \code{MSE}, or a list of \code{Hist} or \code{MSE} objects}

\item{model}{An optional name for the model.
If \code{x} is a list of objects, \code{model} will be taken from \code{names(x)}.
If \code{names(x)} is \code{NULL}, \code{model} will be given sequential numerical values (e.g., \verb{Model 1}, \verb{Model 2}, ...)}
}
\value{
A data.frame
}
\description{
Extracts the life-history parameters: \code{Linf}, \code{K}, \code{L50}, and \code{ageM}
}
