% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_frames.R
\name{get_Years}
\alias{get_Years}
\alias{get_Years.MSE}
\alias{get_Years.MMSE}
\alias{get_Years.Hist}
\alias{get_Years.multiHist}
\title{Create a data.frame with Historical and Projection years}
\usage{
get_Years(x)

\method{get_Years}{MSE}(x)

\method{get_Years}{MMSE}(x)

\method{get_Years}{Hist}(x)

\method{get_Years}{multiHist}(x)
}
\arguments{
\item{x}{An object of class \code{Hist}, \code{MSE}, or a list of \code{Hist} or \code{MSE} objects}
}
\value{
A data.frame with years and period (Historical or Projection)
}
\description{
Create a data.frame with Historical and Projection years
}
