% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{TS_Variables}
\alias{TS_Variables}
\title{Time Series Variables}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
TS_Variables
}
\description{
Time Series Variables
}
\keyword{datasets}
