% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{userguide}
\alias{userguide}
\title{Open the openMSE Documentation website}
\usage{
userguide()
}
\value{
Nothing is returned. Opens the 'openMSE.com' in the web browser
}
\description{
Opens the openMSE Documentation website (requires internet connection)
}
\examples{
userguide()
}
